refactor: split divine names data by catagory
