+++ lib/names/divine-names.ts
@@ -1,1131 +1,9 @@
-export type NameCategory = "dhat" | "sifat" | "af'al";
+import { DIVINE_NAMES } from "./divine-names/data";
+import type { DivineName, NameCategory } from "./divine-names/types";
 
-export interface DivineName {
-  id: number;
-  slug: string;
-  arabic: string;
-  transliteration: string;
-  meaning: string;
-  category: NameCategory;
-  root: string;
-  description: string;
-}
-
-export const CATEGORY_LABELS: Record<
-  NameCategory,
-  { en: string; ar: string; description: string }
-> = {
-  dhat: {
-    en: "Sifat al-Dhat",
-    ar: "صفات الذات",
-    description:
-      "Attributes of the Divine Essence — necessarily and eternally true of Allah independent of creation",
-  },
-  sifat: {
-    en: "Sifat al-Ma'ani",
-    ar: "صفات المعاني",
-    description:
-      "Attributes of Meaning — the seven positive qualities (Knowledge, Power, Will, Life, Hearing, Sight, Speech) and names expressing tanzih (transcendence)",
-  },
-  "af'al": {
-    en: "Sifat al-Af'al",
-    ar: "صفات الأفعال",
-    description:
-      "Attributes of Act — names that express Allah's sovereign relation to creation: creating, providing, guiding, forgiving",
-  },
-};
-
-export const DIVINE_NAMES: DivineName[] = [
-  {
-    id: 1,
-    slug: "ar-rahman",
-    arabic: "الرَّحْمَٰن",
-    transliteration: "Ar-Rahmān",
-    meaning: "The Most Gracious",
-    category: "sifat",
-    root: "ر-ح-م",
-    description:
-      "The all-encompassing mercy that pervades existence; a mercy whose vastness is prior to any creature's merit or request.",
-  },
-  {
-    id: 2,
-    slug: "ar-rahim",
-    arabic: "الرَّحِيم",
-    transliteration: "Ar-Rahīm",
-    meaning: "The Most Merciful",
-    category: "sifat",
-    root: "ر-ح-م",
-    description:
-      "Mercy directed specifically and actively toward the believers; the realised bestowal flowing from the divine attribute of Rahma.",
-  },
-  {
-    id: 3,
-    slug: "al-malik",
-    arabic: "الْمَلِك",
-    transliteration: "Al-Malik",
-    meaning: "The Sovereign",
-    category: "dhat",
-    root: "م-ل-ك",
-    description:
-      "Absolute dominion over all that exists, unconstrained by any partner, predecessor, or limitation — sovereignty that belongs to the Essence itself.",
-  },
-  {
-    id: 4,
-    slug: "al-quddus",
-    arabic: "الْقُدُّوس",
-    transliteration: "Al-Quddūs",
-    meaning: "The Most Holy",
-    category: "sifat",
-    root: "ق-د-س",
-    description:
-      "Perfect freedom from every imperfection, deficiency, or resemblance to created things; the Maturidi tanzih par excellence.",
-  },
-  {
-    id: 5,
-    slug: "as-salam",
-    arabic: "السَّلَام",
-    transliteration: "As-Salām",
-    meaning: "The Source of Peace",
-    category: "sifat",
-    root: "س-ل-م",
-    description:
-      "Allah is free from all defect and the source from which every peace and soundness in creation derives.",
-  },
-  {
-    id: 6,
-    slug: "al-mumin",
-    arabic: "الْمُؤْمِن",
-    transliteration: "Al-Mu'min",
-    meaning: "The Guardian of Faith",
-    category: "af'al",
-    root: "أ-م-ن",
-    description:
-      "He who grants security to His creation and ratifies His own promise — the Originator of all trust and safety.",
-  },
-  {
-    id: 7,
-    slug: "al-muhaymin",
-    arabic: "الْمُهَيْمِن",
-    transliteration: "Al-Muhaymin",
-    meaning: "The Overseer",
-    category: "af'al",
-    root: "ه-ي-م",
-    description:
-      "The all-watchful guardian who holds sovereign oversight over every affair — nothing transpires outside His encompassing awareness.",
-  },
-  {
-    id: 8,
-    slug: "al-aziz",
-    arabic: "الْعَزِيز",
-    transliteration: "Al-'Azīz",
-    meaning: "The Almighty",
-    category: "sifat",
-    root: "ع-ز-ز",
-    description:
-      "Irresistible might that cannot be overcome; the uniqueness of power that has no peer or equal in the divine or created realms.",
-  },
-  {
-    id: 9,
-    slug: "al-jabbar",
-    arabic: "الْجَبَّار",
-    transliteration: "Al-Jabbār",
-    meaning: "The Compeller",
-    category: "sifat",
-    root: "ج-ب-ر",
-    description:
-      "He who repairs what is broken in creation and compels all things according to His will; the overwhelming restorer whose decree none can repel.",
-  },
-  {
-    id: 10,
-    slug: "al-mutakabbir",
-    arabic: "الْمُتَكَبِّر",
-    transliteration: "Al-Mutakabbir",
-    meaning: "The Supreme in Greatness",
-    category: "sifat",
-    root: "ك-ب-ر",
-    description:
-      "Greatness that belongs exclusively to Allah; all created greatness is borrowed dignity, while His is the attribute of the Essence.",
-  },
-  {
-    id: 11,
-    slug: "al-khaliq",
-    arabic: "الْخَالِق",
-    transliteration: "Al-Khāliq",
-    meaning: "The Creator",
-    category: "af'al",
-    root: "خ-ل-ق",
-    description:
-      "The sole originator who brings things into existence by will and command — creation belongs exclusively to His act, not to any secondary cause.",
-  },
-  {
-    id: 12,
-    slug: "al-bari",
-    arabic: "الْبَارِئ",
-    transliteration: "Al-Bāri'",
-    meaning: "The Originator",
-    category: "af'al",
-    root: "ب-ر-أ",
-    description:
-      "He who distinguishes and separates created things from one another, assigning to each its own nature and form.",
-  },
-  {
-    id: 13,
-    slug: "al-musawwir",
-    arabic: "الْمُصَوِّر",
-    transliteration: "Al-Musawwir",
-    meaning: "The Fashioner of Forms",
-    category: "af'al",
-    root: "ص-و-ر",
-    description:
-      "He who gives every created thing its unique form and image according to a wisdom that transcends created imagination.",
-  },
-  {
-    id: 14,
-    slug: "al-ghaffar",
-    arabic: "الْغَفَّار",
-    transliteration: "Al-Ghaffār",
-    meaning: "The All-Forgiving",
-    category: "af'al",
-    root: "غ-ف-ر",
-    description:
-      "The one who repeatedly and abundantly covers sins with His forgiveness — a forgiveness that is an act of pure grace without diminishing His justice.",
-  },
-  {
-    id: 15,
-    slug: "al-qahhar",
-    arabic: "الْقَهَّار",
-    transliteration: "Al-Qahhār",
-    meaning: "The All-Subduing",
-    category: "sifat",
-    root: "ق-ه-ر",
-    description:
-      "He before whose power all things are utterly subdued; no king, force, or angel has any independence before His overwhelming dominion.",
-  },
-  {
-    id: 16,
-    slug: "al-wahhab",
-    arabic: "الْوَهَّاب",
-    transliteration: "Al-Wahhāb",
-    meaning: "The Bestower",
-    category: "af'al",
-    root: "و-ه-ب",
-    description:
-      "The one who gives abundantly without any expectation of return — His bestowal is pure generosity from an essence that has no need.",
-  },
-  {
-    id: 17,
-    slug: "ar-razzaq",
-    arabic: "الرَّزَّاق",
-    transliteration: "Ar-Razzāq",
-    meaning: "The All-Provider",
-    category: "af'al",
-    root: "ر-ز-ق",
-    description:
-      "The ultimate sustainer of all created things — both outer provision of sustenance and inner provision of spiritual capacity.",
-  },
-  {
-    id: 18,
-    slug: "al-fattah",
-    arabic: "الْفَتَّاح",
-    transliteration: "Al-Fattāh",
-    meaning: "The Opener",
-    category: "af'al",
-    root: "ف-ت-ح",
-    description:
-      "The one who opens closed matters, grants victory, and decrees judgment between the creation — all doors of mercy and provision open by His leave alone.",
-  },
-  {
-    id: 19,
-    slug: "al-alim",
-    arabic: "الْعَلِيم",
-    transliteration: "Al-'Alīm",
-    meaning: "The All-Knowing",
-    category: "sifat",
-    root: "ع-ل-م",
-    description:
-      "Infinite knowledge encompassing all particulars and universals, past and future, manifest and hidden — one of the seven essential Sifat al-Ma'ani in Maturidi theology.",
-  },
-  {
-    id: 20,
-    slug: "al-qabid",
-    arabic: "الْقَابِض",
-    transliteration: "Al-Qābid",
-    meaning: "The Withholder",
-    category: "af'al",
-    root: "ق-ب-ض",
-    description:
-      "He who constricts and withholds according to wisdom — trials, scarcity, and contraction of the heart are acts that serve the greater divine order.",
-  },
-  {
-    id: 21,
-    slug: "al-basit",
-    arabic: "الْبَاسِط",
-    transliteration: "Al-Bāsit",
-    meaning: "The Extender",
-    category: "af'al",
-    root: "ب-س-ط",
-    description:
-      "He who expands provision, ease, and the heart's capacity — always paired with Al-Qabid so that expansion and contraction reveal one sovereign wisdom.",
-  },
-  {
-    id: 22,
-    slug: "al-khafid",
-    arabic: "الْخَافِض",
-    transliteration: "Al-Khāfid",
-    meaning: "The Abaser",
-    category: "af'al",
-    root: "خ-ف-ض",
-    description:
-      "He who lowers the arrogant and the unjust; the one in whose hand every station rises or falls according to divine wisdom and justice.",
-  },
-  {
-    id: 23,
-    slug: "ar-rafi",
-    arabic: "الرَّافِع",
-    transliteration: "Ar-Rāfi'",
-    meaning: "The Exalter",
-    category: "af'al",
-    root: "ر-ف-ع",
-    description:
-      "He who elevates whomever He wills among His creation — raising the believers in this life and the next as an act of pure divine favour.",
-  },
-  {
-    id: 24,
-    slug: "al-muizz",
-    arabic: "الْمُعِزّ",
-    transliteration: "Al-Mu'izz",
-    meaning: "The Bestower of Honour",
-    category: "af'al",
-    root: "ع-ز-ز",
-    description:
-      "He who grants might and dignity to whomever He chooses — all true honour in creation is a reflection of His bestowal, not an intrinsic quality of the creature.",
-  },
-  {
-    id: 25,
-    slug: "al-mudhill",
-    arabic: "الْمُذِلّ",
-    transliteration: "Al-Mudhill",
-    meaning: "The Humiliator",
-    category: "af'al",
-    root: "ذ-ل-ل",
-    description:
-      "He who brings low the rebellious and the proud — a reminder that no created greatness is self-sustaining, and all dignity depends on divine permission.",
-  },
-  {
-    id: 26,
-    slug: "as-sami",
-    arabic: "السَّمِيع",
-    transliteration: "As-Samī'",
-    meaning: "The All-Hearing",
-    category: "sifat",
-    root: "س-م-ع",
-    description:
-      "Hearing that encompasses every sound, whisper, and secret thought — one of the seven essential Sifat al-Ma'ani; affirmed without likening it to created hearing.",
-  },
-  {
-    id: 27,
-    slug: "al-basir",
-    arabic: "الْبَصِير",
-    transliteration: "Al-Basīr",
-    meaning: "The All-Seeing",
-    category: "sifat",
-    root: "ب-ص-ر",
-    description:
-      "Vision that encompasses all things simultaneously — one of the seven Sifat al-Ma'ani; affirmed in a manner befitting divine majesty, free from analogy with created sight.",
-  },
-  {
-    id: 28,
-    slug: "al-hakam",
-    arabic: "الْحَكَم",
-    transliteration: "Al-Hakam",
-    meaning: "The Arbitrator",
-    category: "sifat",
-    root: "ح-ك-م",
-    description:
-      "The final judge whose ruling is always just and definitive — no verdict in creation is truly binding except by reference to His judgment.",
-  },
-  {
-    id: 29,
-    slug: "al-adl",
-    arabic: "الْعَدْل",
-    transliteration: "Al-'Adl",
-    meaning: "The Just",
-    category: "sifat",
-    root: "ع-د-ل",
-    description:
-      "Absolute justice that gives every thing its due — the Maturidi school affirms divine justice while holding that Allah is not constrained by creaturely notions of obligation.",
-  },
-  {
-    id: 30,
-    slug: "al-latif",
-    arabic: "اللَّطِيف",
-    transliteration: "Al-Latīf",
-    meaning: "The Subtle",
-    category: "sifat",
-    root: "ل-ط-ف",
-    description:
-      "Awareness so fine and care so gentle that it reaches what no created perception can access — the hidden providences that weave through every life.",
-  },
-  {
-    id: 31,
-    slug: "al-khabir",
-    arabic: "الْخَبِير",
-    transliteration: "Al-Khabīr",
-    meaning: "The All-Aware",
-    category: "sifat",
-    root: "خ-ب-ر",
-    description:
-      "Thorough awareness of the inner realities and hidden dimensions of all things — knowledge of what lies beneath appearances.",
-  },
-  {
-    id: 32,
-    slug: "al-halim",
-    arabic: "الْحَلِيم",
-    transliteration: "Al-Halīm",
-    meaning: "The Forbearing",
-    category: "sifat",
-    root: "ح-ل-م",
-    description:
-      "He who withholds punishment despite the capacity to act and the cause to do so — forebearance that is an expression of wisdom, not weakness.",
-  },
-  {
-    id: 33,
-    slug: "al-azim",
-    arabic: "الْعَظِيم",
-    transliteration: "Al-'Azīm",
-    meaning: "The Magnificent",
-    category: "dhat",
-    root: "ع-ظ-م",
-    description:
-      "A greatness that is beyond all comprehension and comparison — the intellect cannot bound it because it is an attribute of the unlimited Divine Essence.",
-  },
-  {
-    id: 34,
-    slug: "al-ghafur",
-    arabic: "الْغَفُور",
-    transliteration: "Al-Ghafūr",
-    meaning: "The Forgiving",
-    category: "af'al",
-    root: "غ-ف-ر",
-    description:
-      "The depth of forgiveness that completely covers and conceals sins — bestowed as an act of His free will and mercy, not as a debt owed to the sinner.",
-  },
-  {
-    id: 35,
-    slug: "ash-shakur",
-    arabic: "الشَّكُور",
-    transliteration: "Ash-Shakūr",
-    meaning: "The Most Appreciative",
-    category: "af'al",
-    root: "ش-ك-ر",
-    description:
-      "He who multiplies the reward for small deeds with unbounded generosity — gratitude from the Divine that reflects the limitless nature of His giving.",
-  },
-  {
-    id: 36,
-    slug: "al-ali",
-    arabic: "الْعَلِيّ",
-    transliteration: "Al-'Alī",
-    meaning: "The Most High",
-    category: "dhat",
-    root: "ع-ل-و",
-    description:
-      "Transcendent elevation above every limitation and comparison — height that is of the Essence, not spatial or positional.",
-  },
-  {
-    id: 37,
-    slug: "al-kabir",
-    arabic: "الْكَبِير",
-    transliteration: "Al-Kabīr",
-    meaning: "The Most Great",
-    category: "dhat",
-    root: "ك-ب-ر",
-    description:
-      "Greatness that surpasses all measurement; every created greatness is finite and contingent while His is absolute and essential.",
-  },
-  {
-    id: 38,
-    slug: "al-hafiz",
-    arabic: "الْحَفِيظ",
-    transliteration: "Al-Hafīz",
-    meaning: "The Preserver",
-    category: "af'al",
-    root: "ح-ف-ظ",
-    description:
-      "He who guards all things from loss and preserves the records of every deed — the entire created order is held in existence only through His preserving act.",
-  },
-  {
-    id: 39,
-    slug: "al-muqit",
-    arabic: "الْمُقِيت",
-    transliteration: "Al-Muqīt",
-    meaning: "The Sustainer",
-    category: "af'al",
-    root: "ق-و-ت",
-    description:
-      "He who sustains and nourishes every creature with exactly what it needs — combining perfect knowledge of need with perfect power to fulfil it.",
-  },
-  {
-    id: 40,
-    slug: "al-hasib",
-    arabic: "الْحَسِيب",
-    transliteration: "Al-Hasīb",
-    meaning: "The Reckoner",
-    category: "af'al",
-    root: "ح-س-ب",
-    description:
-      "The all-sufficient reckoner who takes precise account of every atom's weight of deed — the one to whom all returns for final accountability.",
-  },
-  {
-    id: 41,
-    slug: "al-jalil",
-    arabic: "الْجَلِيل",
-    transliteration: "Al-Jalīl",
-    meaning: "The Majestic",
-    category: "dhat",
-    root: "ج-ل-ل",
-    description:
-      "Majesty that generates awe and reverence; the jalal aspect of divinity that is balanced in Islamic spirituality by jamal (beauty).",
-  },
-  {
-    id: 42,
-    slug: "al-karim",
-    arabic: "الْكَرِيم",
-    transliteration: "Al-Karīm",
-    meaning: "The Most Generous",
-    category: "af'al",
-    root: "ك-ر-م",
-    description:
-      "Generosity so complete that He gives without need, withholds without injustice, and exceeds every expectation with His gifts.",
-  },
-  {
-    id: 43,
-    slug: "ar-raqib",
-    arabic: "الرَّقِيب",
-    transliteration: "Ar-Raqīb",
-    meaning: "The Watchful",
-    category: "sifat",
-    root: "ر-ق-ب",
-    description:
-      "The ever-vigilant witness who observes every intention and act — a divine attribute that is both protection for the sincere and warning for the heedless.",
-  },
-  {
-    id: 44,
-    slug: "al-mujib",
-    arabic: "الْمُجِيب",
-    transliteration: "Al-Mujīb",
-    meaning: "The Responsive",
-    category: "af'al",
-    root: "ج-و-ب",
-    description:
-      "He who responds to every supplication — the response may be immediate fulfilment, a better alternative, or stored reward, but no sincere call goes unheard.",
-  },
-  {
-    id: 45,
-    slug: "al-wasi",
-    arabic: "الْوَاسِع",
-    transliteration: "Al-Wāsi'",
-    meaning: "The All-Encompassing",
-    category: "dhat",
-    root: "و-س-ع",
-    description:
-      "Boundless in knowledge, mercy, and generosity — an attribute of the Essence that has no border or limit that any created reality can approach.",
-  },
-  {
-    id: 46,
-    slug: "al-hakim",
-    arabic: "الْحَكِيم",
-    transliteration: "Al-Hakīm",
-    meaning: "The All-Wise",
-    category: "sifat",
-    root: "ح-ك-م",
-    description:
-      "Wisdom that places everything in its perfect position — every divine act, decree, and law proceeds from an infinite wisdom that the created intellect can only partially glimpse.",
-  },
-  {
-    id: 47,
-    slug: "al-wadud",
-    arabic: "الْوَدُود",
-    transliteration: "Al-Wadūd",
-    meaning: "The Most Loving",
-    category: "sifat",
-    root: "و-د-د",
-    description:
-      "Divine love that is the origin of all affection in creation — a love that initiates and does not merely respond, freely given to the sincere servant.",
-  },
-  {
-    id: 48,
-    slug: "al-majid",
-    arabic: "الْمَجِيد",
-    transliteration: "Al-Majīd",
-    meaning: "The Most Glorious",
-    category: "dhat",
-    root: "م-ج-د",
-    description:
-      "Glory combining greatness with generosity — a majesty that is not remote or forbidding but radiates outward in gifts and honours upon creation.",
-  },
-  {
-    id: 49,
-    slug: "al-baith",
-    arabic: "الْبَاعِث",
-    transliteration: "Al-Bā'ith",
-    meaning: "The Resurrector",
-    category: "af'al",
-    root: "ب-ع-ث",
-    description:
-      "He who raises the dead and sends prophets — the power of origination reasserted: as He created the first time, so He recreates without the slightest difficulty.",
-  },
-  {
-    id: 50,
-    slug: "ash-shahid",
-    arabic: "الشَّهِيد",
-    transliteration: "Ash-Shahīd",
-    meaning: "The Witness",
-    category: "sifat",
-    root: "ش-ه-د",
-    description:
-      "The all-witnessing presence whose testimony is the ultimate ground of truth — nothing is absent from His witnessing, whether in the visible or unseen realms.",
-  },
-  {
-    id: 51,
-    slug: "al-haqq",
-    arabic: "الْحَقّ",
-    transliteration: "Al-Haqq",
-    meaning: "The Truth",
-    category: "dhat",
-    root: "ح-ق-ق",
-    description:
-      "The necessary, self-subsisting reality — the only being whose existence is not contingent; all else is real only insofar as it participates in His truth.",
-  },
-  {
-    id: 52,
-    slug: "al-wakil",
-    arabic: "الْوَكِيل",
-    transliteration: "Al-Wakīl",
-    meaning: "The Trustee",
-    category: "af'al",
-    root: "و-ك-ل",
-    description:
-      "The disposer of affairs for all who entrust themselves to Him — tawakkul (reliance) is the creature's response to this divine name.",
-  },
-  {
-    id: 53,
-    slug: "al-qawi",
-    arabic: "الْقَوِيّ",
-    transliteration: "Al-Qawī",
-    meaning: "The All-Powerful",
-    category: "sifat",
-    root: "ق-و-ي",
-    description:
-      "Strength that knows no fatigue, depletion, or diminishment — the power of the Essence itself, which no resistance can affect.",
-  },
-  {
-    id: 54,
-    slug: "al-matin",
-    arabic: "الْمَتِين",
-    transliteration: "Al-Matīn",
-    meaning: "The Firm",
-    category: "sifat",
-    root: "م-ت-ن",
-    description:
-      "Absolute firmness in power and resolution — divine will and capacity are unshakeable, and no created force can cause the slightest wavering in His decree.",
-  },
-  {
-    id: 55,
-    slug: "al-wali",
-    arabic: "الْوَلِيّ",
-    transliteration: "Al-Walī",
-    meaning: "The Protecting Friend",
-    category: "af'al",
-    root: "و-ل-ي",
-    description:
-      "The patron and ally who takes the believer under His guardianship — divine friendship that transforms the life of the one who draws near.",
-  },
-  {
-    id: 56,
-    slug: "al-hamid",
-    arabic: "الْحَمِيد",
-    transliteration: "Al-Hamīd",
-    meaning: "The All-Praiseworthy",
-    category: "dhat",
-    root: "ح-م-د",
-    description:
-      "All praise in creation flows toward Him as to its natural source — praise that is not earned but belongs to the Essence as a necessary attribute.",
-  },
-  {
-    id: 57,
-    slug: "al-muhsi",
-    arabic: "الْمُحْصِي",
-    transliteration: "Al-Muhsī",
-    meaning: "The All-Enumerating",
-    category: "sifat",
-    root: "ح-ص-ي",
-    description:
-      "The knowledge that counts and records every particular without the slightest omission — not a leaf falls, nor a thought arises, outside this precise divine enumeration.",
-  },
-  {
-    id: 58,
-    slug: "al-mubdi",
-    arabic: "الْمُبْدِئ",
-    transliteration: "Al-Mubdi'",
-    meaning: "The Originator",
-    category: "af'al",
-    root: "ب-د-أ",
-    description:
-      "He who begins creation without precedent or model — the initial act of bringing forth that demonstrates unlimited creative power.",
-  },
-  {
-    id: 59,
-    slug: "al-muid",
-    arabic: "الْمُعِيد",
-    transliteration: "Al-Mu'īd",
-    meaning: "The Restorer",
-    category: "af'al",
-    root: "ع-و-د",
-    description:
-      "He who returns creation to Himself after its passage through time — the resurrection and return that is the completion of the divine originating act.",
-  },
-  {
-    id: 60,
-    slug: "al-muhyi",
-    arabic: "الْمُحْيِي",
-    transliteration: "Al-Muhyī",
-    meaning: "The Giver of Life",
-    category: "af'al",
-    root: "ح-ي-ي",
-    description:
-      "The sole originator of life in all its forms — physical life, spiritual vitality, and the life of the resurrection all belong to His act alone.",
-  },
-  {
-    id: 61,
-    slug: "al-mumit",
-    arabic: "الْمُمِيت",
-    transliteration: "Al-Mumīt",
-    meaning: "The Taker of Life",
-    category: "af'al",
-    root: "م-و-ت",
-    description:
-      "He who decrees the death of every living thing — paired with Al-Muhyi to affirm that life and death are entirely in His hand, neither random nor outside His will.",
-  },
-  {
-    id: 62,
-    slug: "al-hayy",
-    arabic: "الْحَيّ",
-    transliteration: "Al-Hayy",
-    meaning: "The Ever-Living",
-    category: "sifat",
-    root: "ح-ي-ي",
-    description:
-      "Life that is not preceded by non-existence and not followed by cessation — the essential divine life that sustains all contingent life; one of the seven Sifat al-Ma'ani.",
-  },
-  {
-    id: 63,
-    slug: "al-qayyum",
-    arabic: "الْقَيُّوم",
-    transliteration: "Al-Qayyūm",
-    meaning: "The Self-Subsisting",
-    category: "dhat",
-    root: "ق-و-م",
-    description:
-      "He who subsists of Himself and on whom all else subsists — the Maturidi attribute of Qiyam bi-l-nafs (self-subsistence) embodied in this name.",
-  },
-  {
-    id: 64,
-    slug: "al-wajid",
-    arabic: "الْوَاجِد",
-    transliteration: "Al-Wājid",
-    meaning: "The Finder",
-    category: "dhat",
-    root: "و-ج-د",
-    description:
-      "He who finds and perceives all things with no searching or seeking — a name of the Essence whose awareness precedes every created event.",
-  },
-  {
-    id: 65,
-    slug: "al-wahid",
-    arabic: "الْوَاحِد",
-    transliteration: "Al-Wāhid",
-    meaning: "The One",
-    category: "dhat",
-    root: "و-ح-د",
-    description:
-      "Oneness in essence, attributes, and acts — the foundation of tawhid; no second being shares in or participates in this oneness.",
-  },
-  {
-    id: 66,
-    slug: "al-ahad",
-    arabic: "الْأَحَد",
-    transliteration: "Al-Ahad",
-    meaning: "The Unique",
-    category: "dhat",
-    root: "و-ح-د",
-    description:
-      "Absolute singularity that admits no composition, division, or partner — the name that Sura Al-Ikhlas uses to negate every conceivable form of association.",
-  },
-  {
-    id: 67,
-    slug: "as-samad",
-    arabic: "الصَّمَد",
-    transliteration: "As-Samad",
-    meaning: "The Eternal",
-    category: "dhat",
-    root: "ص-م-د",
-    description:
-      "The self-sufficient one to whom all creation turns in need — possessing no need of any kind while being the ultimate object of every creature's need.",
-  },
-  {
-    id: 68,
-    slug: "al-qadir",
-    arabic: "الْقَادِر",
-    transliteration: "Al-Qādir",
-    meaning: "The All-Capable",
-    category: "sifat",
-    root: "ق-د-ر",
-    description:
-      "Unlimited power over all possible things — one of the seven Sifat al-Ma'ani; the Maturidi school affirms that divine power does not extend to logical impossibilities, but this is a limit of logic, not of power.",
-  },
-  {
-    id: 69,
-    slug: "al-muqtadir",
-    arabic: "الْمُقْتَدِر",
-    transliteration: "Al-Muqtadir",
-    meaning: "The All-Determining",
-    category: "sifat",
-    root: "ق-د-ر",
-    description:
-      "He who exercises His power with decisive authority — a name that emphasises the active, sovereign exercise of the divine power.",
-  },
-  {
-    id: 70,
-    slug: "al-muqaddim",
-    arabic: "الْمُقَدِّم",
-    transliteration: "Al-Muqaddim",
-    meaning: "The Expediter",
-    category: "af'al",
-    root: "ق-د-م",
-    description:
-      "He who brings forward whomever and whatever He wills — advancing the believers, the prophets, and the causes that serve His purpose.",
-  },
-  {
-    id: 71,
-    slug: "al-muakhkhir",
-    arabic: "الْمُؤَخِّر",
-    transliteration: "Al-Mu'akhkhir",
-    meaning: "The Delayer",
-    category: "af'al",
-    root: "أ-خ-ر",
-    description:
-      "He who postpones and delays according to wisdom — paired with Al-Muqaddim to show that all timing belongs to the divine will, not to chance.",
-  },
-  {
-    id: 72,
-    slug: "al-awwal",
-    arabic: "الْأَوَّل",
-    transliteration: "Al-Awwal",
-    meaning: "The First",
-    category: "dhat",
-    root: "أ-و-ل",
-    description:
-      "He who precedes all things with no beginning — existence prior to time and before any created reality; the Qidam (eternity a parte ante) of the Essence.",
-  },
-  {
-    id: 73,
-    slug: "al-akhir",
-    arabic: "الْآخِر",
-    transliteration: "Al-Ākhir",
-    meaning: "The Last",
-    category: "dhat",
-    root: "أ-خ-ر",
-    description:
-      "He who remains after all things — Baqa (eternal permanence); when all created things pass away, only the Divine Essence persists.",
-  },
-  {
-    id: 74,
-    slug: "az-zahir",
-    arabic: "الظَّاهِر",
-    transliteration: "Az-Zāhir",
-    meaning: "The Manifest",
-    category: "dhat",
-    root: "ظ-ه-ر",
-    description:
-      "He who is manifest through His signs, effects, and clear proofs — more evident than any created thing, though not grasped by direct sensory perception.",
-  },
-  {
-    id: 75,
-    slug: "al-batin",
-    arabic: "الْبَاطِن",
-    transliteration: "Al-Bātin",
-    meaning: "The Hidden",
-    category: "dhat",
-    root: "ب-ط-ن",
-    description:
-      "He who is hidden from direct comprehension — transcendence so complete that the Essence itself is beyond the reach of any created faculty.",
-  },
-  {
-    id: 76,
-    slug: "al-wali-governor",
-    arabic: "الْوَالِي",
-    transliteration: "Al-Wālī",
-    meaning: "The Governor",
-    category: "af'al",
-    root: "و-ل-ي",
-    description:
-      "He who governs and administers the entire created order — the sovereign management of the cosmos that proceeds from His unlimited knowledge and will.",
-  },
-  {
-    id: 77,
-    slug: "al-mutaali",
-    arabic: "الْمُتَعَالِي",
-    transliteration: "Al-Muta'ālī",
-    meaning: "The Most Exalted",
-    category: "sifat",
-    root: "ع-ل-و",
-    description:
-      "He who is supremely exalted above every comparison, attribution, or limitation — the ultimate expression of tanzih (transcendence) in divine names.",
-  },
-  {
-    id: 78,
-    slug: "al-barr",
-    arabic: "الْبَرّ",
-    transliteration: "Al-Barr",
-    meaning: "The Source of Goodness",
-    category: "af'al",
-    root: "ب-ر-ر",
-    description:
-      "The origin of every form of goodness and beneficence — all that is good in creation is a trace of the divine al-Birr radiating outward.",
-  },
-  {
-    id: 79,
-    slug: "at-tawwab",
-    arabic: "التَّوَّاب",
-    transliteration: "At-Tawwāb",
-    meaning: "The Acceptor of Repentance",
-    category: "af'al",
-    root: "ت-و-ب",
-    description:
-      "He who ceaselessly turns toward the repentant with acceptance — the divine turning precedes, enables, and completes the human turning; both the initiator and receiver of tawba.",
-  },
-  {
-    id: 80,
-    slug: "al-muntaqim",
-    arabic: "الْمُنْتَقِم",
-    transliteration: "Al-Muntaqim",
-    meaning: "The Avenger",
-    category: "af'al",
-    root: "ن-ق-م",
-    description:
-      "He who executes justice against the unjust — divine retribution that is not rage but the measured correction of the cosmic moral order.",
-  },
-  {
-    id: 81,
-    slug: "al-afuw",
-    arabic: "الْعَفُوّ",
-    transliteration: "Al-'Afuw",
-    meaning: "The Pardoner",
-    category: "af'al",
-    root: "ع-ف-و",
-    description:
-      "He who erases sins completely — not merely covering them as Al-Ghafur does but effacing them entirely so they leave no trace in the divine record.",
-  },
-  {
-    id: 82,
-    slug: "ar-rauf",
-    arabic: "الرَّؤُوف",
-    transliteration: "Ar-Ra'ūf",
-    meaning: "The Most Compassionate",
-    category: "sifat",
-    root: "ر-أ-ف",
-    description:
-      "A compassion more intense than mercy — the tenderness of the divine care that acts to remove hardship before it overwhelms the beloved creature.",
-  },
-  {
-    id: 83,
-    slug: "malik-ul-mulk",
-    arabic: "مَالِكُ الْمُلْك",
-    transliteration: "Mālik al-Mulk",
-    meaning: "Owner of All Sovereignty",
-    category: "dhat",
-    root: "م-ل-ك",
-    description:
-      "Absolute ownership of every realm, domain, and kingdom — no sovereignty exists independently; all thrones are merely given in trust and retrieved at will.",
-  },
-  {
-    id: 84,
-    slug: "dhul-jalali-wal-ikram",
-    arabic: "ذُو الْجَلَالِ وَالْإِكْرَام",
-    transliteration: "Dhū al-Jalāl wa-l-Ikrām",
-    meaning: "Lord of Majesty and Bounty",
-    category: "dhat",
-    root: "ج-ل-ل",
-    description:
-      "The pairing of transcendent awe-inspiring majesty (jalal) with active honouring generosity (ikram) — the complete divine character toward which all worship is directed.",
-  },
-  {
-    id: 85,
-    slug: "al-muqsit",
-    arabic: "الْمُقْسِط",
-    transliteration: "Al-Muqsit",
-    meaning: "The Equitable",
-    category: "af'al",
-    root: "ق-س-ط",
-    description:
-      "He who deals with perfect fairness — equity in judgment, proportion in reward and recompense, and the source of all just dealings in the social order.",
-  },
-  {
-    id: 86,
-    slug: "al-jami",
-    arabic: "الْجَامِع",
-    transliteration: "Al-Jāmi'",
-    meaning: "The Gatherer",
-    category: "af'al",
-    root: "ج-م-ع",
-    description:
-      "He who gathers all things for the Day of Assembly — the gathering of dispersed realities into one moment of accountability and completion.",
-  },
-  {
-    id: 87,
-    slug: "al-ghani",
-    arabic: "الْغَنِيّ",
-    transliteration: "Al-Ghanī",
-    meaning: "The Self-Sufficient",
-    category: "dhat",
-    root: "غ-ن-ي",
-    description:
-      "Absolute richness and independence — needing nothing from creation while being the source from which all created needs are satisfied.",
-  },
-  {
-    id: 88,
-    slug: "al-mughni",
-    arabic: "الْمُغْنِي",
-    transliteration: "Al-Mughnī",
-    meaning: "The Enricher",
-    category: "af'al",
-    root: "غ-ن-ي",
-    description:
-      "He who removes need and bestows sufficiency upon whomever He wills — enrichment of the body, spirit, and heart all flow from this divine act.",
-  },
-  {
-    id: 89,
-    slug: "al-mani",
-    arabic: "الْمَانِع",
-    transliteration: "Al-Māni'",
-    meaning: "The Preventer",
-    category: "af'al",
-    root: "م-ن-ع",
-    description:
-      "He who withholds what would harm and prevents what is not due — the prevention that is itself a mercy, though the creature may not perceive it as such.",
-  },
-  {
-    id: 90,
-    slug: "ad-darr",
-    arabic: "الضَّارّ",
-    transliteration: "Ad-Dārr",
-    meaning: "The Distresser",
-    category: "af'al",
-    root: "ض-ر-ر",
-    description:
-      "He who sends trial and hardship by His will — in Maturidi thought, harm as well as benefit originates solely from the divine act, affirming tawhid in all circumstances.",
-  },
-  {
-    id: 91,
-    slug: "an-nafi",
-    arabic: "النَّافِع",
-    transliteration: "An-Nāfi'",
-    meaning: "The Benefiter",
-    category: "af'al",
-    root: "ن-ف-ع",
-    description:
-      "He who creates and directs all benefit — paired with Ad-Darr to affirm that both benefit and harm come from the same divine will that is entirely wise and just.",
-  },
-  {
-    id: 92,
-    slug: "an-nur",
-    arabic: "النُّور",
-    transliteration: "An-Nūr",
-    meaning: "The Light",
-    category: "dhat",
-    root: "ن-و-ر",
-    description:
-      "The light of the heavens and the earth — not physical light but the ontological illumination by which anything can be known, seen, or guided.",
-  },
-  {
-    id: 93,
-    slug: "al-hadi",
-    arabic: "الْهَادِي",
-    transliteration: "Al-Hādī",
-    meaning: "The Guide",
-    category: "af'al",
-    root: "ه-د-ي",
-    description:
-      "The sovereign guide who directs whomever He wills to the straight path — divine guidance (tawfiq) is an act from Him, while human choice operates within this framework.",
-  },
-  {
-    id: 94,
-    slug: "al-badi",
-    arabic: "الْبَدِيع",
-    transliteration: "Al-Badī'",
-    meaning: "The Incomparable Originator",
-    category: "af'al",
-    root: "ب-د-ع",
-    description:
-      "He who creates in ways that have no precedent or parallel — every act of creation is unprecedented because it flows from an essence that is itself incomparable.",
-  },
-  {
-    id: 95,
-    slug: "al-baqi",
-    arabic: "الْبَاقِي",
-    transliteration: "Al-Bāqī",
-    meaning: "The Everlasting",
-    category: "dhat",
-    root: "ب-ق-ي",
-    description:
-      "Eternal permanence with no end — the Baqa attribute of the divine Essence; all created existence is borrowed and temporary while His is necessary and permanent.",
-  },
-  {
-    id: 96,
-    slug: "al-warith",
-    arabic: "الْوَارِث",
-    transliteration: "Al-Wārith",
-    meaning: "The Inheritor",
-    category: "af'al",
-    root: "و-ر-ث",
-    description:
-      "He to whom all returns after the passing of every created owner — the final and only enduring possessor when all created claims have expired.",
-  },
-  {
-    id: 97,
-    slug: "ar-rashid",
-    arabic: "الرَّشِيد",
-    transliteration: "Ar-Rashīd",
-    meaning: "The Guide to the Right Path",
-    category: "sifat",
-    root: "ر-ش-د",
-    description:
-      "He whose every act is perfectly well-directed and guided — wisdom so complete that nothing He does is wasteful, misdirected, or improvised.",
-  },
-  {
-    id: 98,
-    slug: "as-sabur",
-    arabic: "الصَّبُور",
-    transliteration: "As-Sabūr",
-    meaning: "The Most Patient",
-    category: "sifat",
-    root: "ص-ب-ر",
-    description:
-      "He who does not hasten punishment despite transgression — divine patience that gives creatures time to return, and that should inspire the believer's own sabr.",
-  },
-  {
-    id: 99,
-    slug: "allah",
-    arabic: "اللَّه",
-    transliteration: "Allāh",
-    meaning: "The God",
-    category: "dhat",
-    root: "أ-ل-ه",
-    description:
-      "The supreme proper name — Al-Ism al-A'zam (the Greatest Name) that encompasses all divine attributes; the name to which all other names are adjectives.",
-  },
-];
+export type { DivineName, NameCategory } from "./divine-names/types";
+export { CATEGORY_LABELS } from "./divine-names/categories";
+export { DIVINE_NAMES } from "./divine-names/data";
 
 export function getNameBySlug(slug: string): DivineName | undefined {
   return DIVINE_NAMES.find((n) => n.slug === slug);
