+++ lib/names/divine-names/data/afal.ts
@@ -1,0 +1,454 @@
+import type { DivineName } from "../types";
+
+export const AFAL_NAMES: DivineName[] = [
+  {
+    id: 6,
+    slug: "al-mumin",
+    arabic: "الْمُؤْمِن",
+    transliteration: "Al-Mu'min",
+    meaning: "The Guardian of Faith",
+    category: "af'al",
+    root: "أ-م-ن",
+    description: "He who grants security to His creation and ratifies His own promise — the Originator of all trust and safety.",
+  },
+  {
+    id: 7,
+    slug: "al-muhaymin",
+    arabic: "الْمُهَيْمِن",
+    transliteration: "Al-Muhaymin",
+    meaning: "The Overseer",
+    category: "af'al",
+    root: "ه-ي-م",
+    description: "The all-watchful guardian who holds sovereign oversight over every affair — nothing transpires outside His encompassing awareness.",
+  },
+  {
+    id: 11,
+    slug: "al-khaliq",
+    arabic: "الْخَالِق",
+    transliteration: "Al-Khāliq",
+    meaning: "The Creator",
+    category: "af'al",
+    root: "خ-ل-ق",
+    description: "The sole originator who brings things into existence by will and command — creation belongs exclusively to His act, not to any secondary cause.",
+  },
+  {
+    id: 12,
+    slug: "al-bari",
+    arabic: "الْبَارِئ",
+    transliteration: "Al-Bāri'",
+    meaning: "The Originator",
+    category: "af'al",
+    root: "ب-ر-أ",
+    description: "He who distinguishes and separates created things from one another, assigning to each its own nature and form.",
+  },
+  {
+    id: 13,
+    slug: "al-musawwir",
+    arabic: "الْمُصَوِّر",
+    transliteration: "Al-Musawwir",
+    meaning: "The Fashioner of Forms",
+    category: "af'al",
+    root: "ص-و-ر",
+    description: "He who gives every created thing its unique form and image according to a wisdom that transcends created imagination.",
+  },
+  {
+    id: 14,
+    slug: "al-ghaffar",
+    arabic: "الْغَفَّار",
+    transliteration: "Al-Ghaffār",
+    meaning: "The All-Forgiving",
+    category: "af'al",
+    root: "غ-ف-ر",
+    description: "The one who repeatedly and abundantly covers sins with His forgiveness — a forgiveness that is an act of pure grace without diminishing His justice.",
+  },
+  {
+    id: 16,
+    slug: "al-wahhab",
+    arabic: "الْوَهَّاب",
+    transliteration: "Al-Wahhāb",
+    meaning: "The Bestower",
+    category: "af'al",
+    root: "و-ه-ب",
+    description: "The one who gives abundantly without any expectation of return — His bestowal is pure generosity from an essence that has no need.",
+  },
+  {
+    id: 17,
+    slug: "ar-razzaq",
+    arabic: "الرَّزَّاق",
+    transliteration: "Ar-Razzāq",
+    meaning: "The All-Provider",
+    category: "af'al",
+    root: "ر-ز-ق",
+    description: "The ultimate sustainer of all created things — both outer provision of sustenance and inner provision of spiritual capacity.",
+  },
+  {
+    id: 18,
+    slug: "al-fattah",
+    arabic: "الْفَتَّاح",
+    transliteration: "Al-Fattāh",
+    meaning: "The Opener",
+    category: "af'al",
+    root: "ف-ت-ح",
+    description: "The one who opens closed matters, grants victory, and decrees judgment between the creation — all doors of mercy and provision open by His leave alone.",
+  },
+  {
+    id: 20,
+    slug: "al-qabid",
+    arabic: "الْقَابِض",
+    transliteration: "Al-Qābid",
+    meaning: "The Withholder",
+    category: "af'al",
+    root: "ق-ب-ض",
+    description: "He who constricts and withholds according to wisdom — trials, scarcity, and contraction of the heart are acts that serve the greater divine order.",
+  },
+  {
+    id: 21,
+    slug: "al-basit",
+    arabic: "الْبَاسِط",
+    transliteration: "Al-Bāsit",
+    meaning: "The Extender",
+    category: "af'al",
+    root: "ب-س-ط",
+    description: "He who expands provision, ease, and the heart's capacity — always paired with Al-Qabid so that expansion and contraction reveal one sovereign wisdom.",
+  },
+  {
+    id: 22,
+    slug: "al-khafid",
+    arabic: "الْخَافِض",
+    transliteration: "Al-Khāfid",
+    meaning: "The Abaser",
+    category: "af'al",
+    root: "خ-ف-ض",
+    description: "He who lowers the arrogant and the unjust; the one in whose hand every station rises or falls according to divine wisdom and justice.",
+  },
+  {
+    id: 23,
+    slug: "ar-rafi",
+    arabic: "الرَّافِع",
+    transliteration: "Ar-Rāfi'",
+    meaning: "The Exalter",
+    category: "af'al",
+    root: "ر-ف-ع",
+    description: "He who elevates whomever He wills among His creation — raising the believers in this life and the next as an act of pure divine favour.",
+  },
+  {
+    id: 24,
+    slug: "al-muizz",
+    arabic: "الْمُعِزّ",
+    transliteration: "Al-Mu'izz",
+    meaning: "The Bestower of Honour",
+    category: "af'al",
+    root: "ع-ز-ز",
+    description: "He who grants might and dignity to whomever He chooses — all true honour in creation is a reflection of His bestowal, not an intrinsic quality of the creature.",
+  },
+  {
+    id: 25,
+    slug: "al-mudhill",
+    arabic: "الْمُذِلّ",
+    transliteration: "Al-Mudhill",
+    meaning: "The Humiliator",
+    category: "af'al",
+    root: "ذ-ل-ل",
+    description: "He who brings low the rebellious and the proud — a reminder that no created greatness is self-sustaining, and all dignity depends on divine permission.",
+  },
+  {
+    id: 34,
+    slug: "al-ghafur",
+    arabic: "الْغَفُور",
+    transliteration: "Al-Ghafūr",
+    meaning: "The Forgiving",
+    category: "af'al",
+    root: "غ-ف-ر",
+    description: "The depth of forgiveness that completely covers and conceals sins — bestowed as an act of His free will and mercy, not as a debt owed to the sinner.",
+  },
+  {
+    id: 35,
+    slug: "ash-shakur",
+    arabic: "الشَّكُور",
+    transliteration: "Ash-Shakūr",
+    meaning: "The Most Appreciative",
+    category: "af'al",
+    root: "ش-ك-ر",
+    description: "He who multiplies the reward for small deeds with unbounded generosity — gratitude from the Divine that reflects the limitless nature of His giving.",
+  },
+  {
+    id: 38,
+    slug: "al-hafiz",
+    arabic: "الْحَفِيظ",
+    transliteration: "Al-Hafīz",
+    meaning: "The Preserver",
+    category: "af'al",
+    root: "ح-ف-ظ",
+    description: "He who guards all things from loss and preserves the records of every deed — the entire created order is held in existence only through His preserving act.",
+  },
+  {
+    id: 39,
+    slug: "al-muqit",
+    arabic: "الْمُقِيت",
+    transliteration: "Al-Muqīt",
+    meaning: "The Sustainer",
+    category: "af'al",
+    root: "ق-و-ت",
+    description: "He who sustains and nourishes every creature with exactly what it needs — combining perfect knowledge of need with perfect power to fulfil it.",
+  },
+  {
+    id: 40,
+    slug: "al-hasib",
+    arabic: "الْحَسِيب",
+    transliteration: "Al-Hasīb",
+    meaning: "The Reckoner",
+    category: "af'al",
+    root: "ح-س-ب",
+    description: "The all-sufficient reckoner who takes precise account of every atom's weight of deed — the one to whom all returns for final accountability.",
+  },
+  {
+    id: 42,
+    slug: "al-karim",
+    arabic: "الْكَرِيم",
+    transliteration: "Al-Karīm",
+    meaning: "The Most Generous",
+    category: "af'al",
+    root: "ك-ر-م",
+    description: "Generosity so complete that He gives without need, withholds without injustice, and exceeds every expectation with His gifts.",
+  },
+  {
+    id: 44,
+    slug: "al-mujib",
+    arabic: "الْمُجِيب",
+    transliteration: "Al-Mujīb",
+    meaning: "The Responsive",
+    category: "af'al",
+    root: "ج-و-ب",
+    description: "He who responds to every supplication — the response may be immediate fulfilment, a better alternative, or stored reward, but no sincere call goes unheard.",
+  },
+  {
+    id: 49,
+    slug: "al-baith",
+    arabic: "الْبَاعِث",
+    transliteration: "Al-Bā'ith",
+    meaning: "The Resurrector",
+    category: "af'al",
+    root: "ب-ع-ث",
+    description: "He who raises the dead and sends prophets — the power of origination reasserted: as He created the first time, so He recreates without the slightest difficulty.",
+  },
+  {
+    id: 52,
+    slug: "al-wakil",
+    arabic: "الْوَكِيل",
+    transliteration: "Al-Wakīl",
+    meaning: "The Trustee",
+    category: "af'al",
+    root: "و-ك-ل",
+    description: "The disposer of affairs for all who entrust themselves to Him — tawakkul (reliance) is the creature's response to this divine name.",
+  },
+  {
+    id: 55,
+    slug: "al-wali",
+    arabic: "الْوَلِيّ",
+    transliteration: "Al-Walī",
+    meaning: "The Protecting Friend",
+    category: "af'al",
+    root: "و-ل-ي",
+    description: "The patron and ally who takes the believer under His guardianship — divine friendship that transforms the life of the one who draws near.",
+  },
+  {
+    id: 58,
+    slug: "al-mubdi",
+    arabic: "الْمُبْدِئ",
+    transliteration: "Al-Mubdi'",
+    meaning: "The Originator",
+    category: "af'al",
+    root: "ب-د-أ",
+    description: "He who begins creation without precedent or model — the initial act of bringing forth that demonstrates unlimited creative power.",
+  },
+  {
+    id: 59,
+    slug: "al-muid",
+    arabic: "الْمُعِيد",
+    transliteration: "Al-Mu'īd",
+    meaning: "The Restorer",
+    category: "af'al",
+    root: "ع-و-د",
+    description: "He who returns creation to Himself after its passage through time — the resurrection and return that is the completion of the divine originating act.",
+  },
+  {
+    id: 60,
+    slug: "al-muhyi",
+    arabic: "الْمُحْيِي",
+    transliteration: "Al-Muhyī",
+    meaning: "The Giver of Life",
+    category: "af'al",
+    root: "ح-ي-ي",
+    description: "The sole originator of life in all its forms — physical life, spiritual vitality, and the life of the resurrection all belong to His act alone.",
+  },
+  {
+    id: 61,
+    slug: "al-mumit",
+    arabic: "الْمُمِيت",
+    transliteration: "Al-Mumīt",
+    meaning: "The Taker of Life",
+    category: "af'al",
+    root: "م-و-ت",
+    description: "He who decrees the death of every living thing — paired with Al-Muhyi to affirm that life and death are entirely in His hand, neither random nor outside His will.",
+  },
+  {
+    id: 70,
+    slug: "al-muqaddim",
+    arabic: "الْمُقَدِّم",
+    transliteration: "Al-Muqaddim",
+    meaning: "The Expediter",
+    category: "af'al",
+    root: "ق-د-م",
+    description: "He who brings forward whomever and whatever He wills — advancing the believers, the prophets, and the causes that serve His purpose.",
+  },
+  {
+    id: 71,
+    slug: "al-muakhkhir",
+    arabic: "الْمُؤَخِّر",
+    transliteration: "Al-Mu'akhkhir",
+    meaning: "The Delayer",
+    category: "af'al",
+    root: "أ-خ-ر",
+    description: "He who postpones and delays according to wisdom — paired with Al-Muqaddim to show that all timing belongs to the divine will, not to chance.",
+  },
+  {
+    id: 76,
+    slug: "al-wali-governor",
+    arabic: "الْوَالِي",
+    transliteration: "Al-Wālī",
+    meaning: "The Governor",
+    category: "af'al",
+    root: "و-ل-ي",
+    description: "He who governs and administers the entire created order — the sovereign management of the cosmos that proceeds from His unlimited knowledge and will.",
+  },
+  {
+    id: 78,
+    slug: "al-barr",
+    arabic: "الْبَرّ",
+    transliteration: "Al-Barr",
+    meaning: "The Source of Goodness",
+    category: "af'al",
+    root: "ب-ر-ر",
+    description: "The origin of every form of goodness and beneficence — all that is good in creation is a trace of the divine al-Birr radiating outward.",
+  },
+  {
+    id: 79,
+    slug: "at-tawwab",
+    arabic: "التَّوَّاب",
+    transliteration: "At-Tawwāb",
+    meaning: "The Acceptor of Repentance",
+    category: "af'al",
+    root: "ت-و-ب",
+    description: "He who ceaselessly turns toward the repentant with acceptance — the divine turning precedes, enables, and completes the human turning; both the initiator and receiver of tawba.",
+  },
+  {
+    id: 80,
+    slug: "al-muntaqim",
+    arabic: "الْمُنْتَقِم",
+    transliteration: "Al-Muntaqim",
+    meaning: "The Avenger",
+    category: "af'al",
+    root: "ن-ق-م",
+    description: "He who executes justice against the unjust — divine retribution that is not rage but the measured correction of the cosmic moral order.",
+  },
+  {
+    id: 81,
+    slug: "al-afuw",
+    arabic: "الْعَفُوّ",
+    transliteration: "Al-'Afuw",
+    meaning: "The Pardoner",
+    category: "af'al",
+    root: "ع-ف-و",
+    description: "He who erases sins completely — not merely covering them as Al-Ghafur does but effacing them entirely so they leave no trace in the divine record.",
+  },
+  {
+    id: 85,
+    slug: "al-muqsit",
+    arabic: "الْمُقْسِط",
+    transliteration: "Al-Muqsit",
+    meaning: "The Equitable",
+    category: "af'al",
+    root: "ق-س-ط",
+    description: "He who deals with perfect fairness — equity in judgment, proportion in reward and recompense, and the source of all just dealings in the social order.",
+  },
+  {
+    id: 86,
+    slug: "al-jami",
+    arabic: "الْجَامِع",
+    transliteration: "Al-Jāmi'",
+    meaning: "The Gatherer",
+    category: "af'al",
+    root: "ج-م-ع",
+    description: "He who gathers all things for the Day of Assembly — the gathering of dispersed realities into one moment of accountability and completion.",
+  },
+  {
+    id: 88,
+    slug: "al-mughni",
+    arabic: "الْمُغْنِي",
+    transliteration: "Al-Mughnī",
+    meaning: "The Enricher",
+    category: "af'al",
+    root: "غ-ن-ي",
+    description: "He who removes need and bestows sufficiency upon whomever He wills — enrichment of the body, spirit, and heart all flow from this divine act.",
+  },
+  {
+    id: 89,
+    slug: "al-mani",
+    arabic: "الْمَانِع",
+    transliteration: "Al-Māni'",
+    meaning: "The Preventer",
+    category: "af'al",
+    root: "م-ن-ع",
+    description: "He who withholds what would harm and prevents what is not due — the prevention that is itself a mercy, though the creature may not perceive it as such.",
+  },
+  {
+    id: 90,
+    slug: "ad-darr",
+    arabic: "الضَّارّ",
+    transliteration: "Ad-Dārr",
+    meaning: "The Distresser",
+    category: "af'al",
+    root: "ض-ر-ر",
+    description: "He who sends trial and hardship by His will — in Maturidi thought, harm as well as benefit originates solely from the divine act, affirming tawhid in all circumstances.",
+  },
+  {
+    id: 91,
+    slug: "an-nafi",
+    arabic: "النَّافِع",
+    transliteration: "An-Nāfi'",
+    meaning: "The Benefiter",
+    category: "af'al",
+    root: "ن-ف-ع",
+    description: "He who creates and directs all benefit — paired with Ad-Darr to affirm that both benefit and harm come from the same divine will that is entirely wise and just.",
+  },
+  {
+    id: 93,
+    slug: "al-hadi",
+    arabic: "الْهَادِي",
+    transliteration: "Al-Hādī",
+    meaning: "The Guide",
+    category: "af'al",
+    root: "ه-د-ي",
+    description: "The sovereign guide who directs whomever He wills to the straight path — divine guidance (tawfiq) is an act from Him, while human choice operates within this framework.",
+  },
+  {
+    id: 94,
+    slug: "al-badi",
+    arabic: "الْبَدِيع",
+    transliteration: "Al-Badī'",
+    meaning: "The Incomparable Originator",
+    category: "af'al",
+    root: "ب-د-ع",
+    description: "He who creates in ways that have no precedent or parallel — every act of creation is unprecedented because it flows from an essence that is itself incomparable.",
+  },
+  {
+    id: 96,
+    slug: "al-warith",
+    arabic: "الْوَارِث",
+    transliteration: "Al-Wārith",
+    meaning: "The Inheritor",
+    category: "af'al",
+    root: "و-ر-ث",
+    description: "He to whom all returns after the passing of every created owner — the final and only enduring possessor when all created claims have expired.",
+  },
+];
