+++ lib/names/divine-names/data/dhat.ts
@@ -1,0 +1,244 @@
+import type { DivineName } from "../types";
+
+export const DHAT_NAMES: DivineName[] = [
+  {
+    id: 3,
+    slug: "al-malik",
+    arabic: "الْمَلِك",
+    transliteration: "Al-Malik",
+    meaning: "The Sovereign",
+    category: "dhat",
+    root: "م-ل-ك",
+    description: "Absolute dominion over all that exists, unconstrained by any partner, predecessor, or limitation — sovereignty that belongs to the Essence itself.",
+  },
+  {
+    id: 33,
+    slug: "al-azim",
+    arabic: "الْعَظِيم",
+    transliteration: "Al-'Azīm",
+    meaning: "The Magnificent",
+    category: "dhat",
+    root: "ع-ظ-م",
+    description: "A greatness that is beyond all comprehension and comparison — the intellect cannot bound it because it is an attribute of the unlimited Divine Essence.",
+  },
+  {
+    id: 36,
+    slug: "al-ali",
+    arabic: "الْعَلِيّ",
+    transliteration: "Al-'Alī",
+    meaning: "The Most High",
+    category: "dhat",
+    root: "ع-ل-و",
+    description: "Transcendent elevation above every limitation and comparison — height that is of the Essence, not spatial or positional.",
+  },
+  {
+    id: 37,
+    slug: "al-kabir",
+    arabic: "الْكَبِير",
+    transliteration: "Al-Kabīr",
+    meaning: "The Most Great",
+    category: "dhat",
+    root: "ك-ب-ر",
+    description: "Greatness that surpasses all measurement; every created greatness is finite and contingent while His is absolute and essential.",
+  },
+  {
+    id: 41,
+    slug: "al-jalil",
+    arabic: "الْجَلِيل",
+    transliteration: "Al-Jalīl",
+    meaning: "The Majestic",
+    category: "dhat",
+    root: "ج-ل-ل",
+    description: "Majesty that generates awe and reverence; the jalal aspect of divinity that is balanced in Islamic spirituality by jamal (beauty).",
+  },
+  {
+    id: 45,
+    slug: "al-wasi",
+    arabic: "الْوَاسِع",
+    transliteration: "Al-Wāsi'",
+    meaning: "The All-Encompassing",
+    category: "dhat",
+    root: "و-س-ع",
+    description: "Boundless in knowledge, mercy, and generosity — an attribute of the Essence that has no border or limit that any created reality can approach.",
+  },
+  {
+    id: 48,
+    slug: "al-majid",
+    arabic: "الْمَجِيد",
+    transliteration: "Al-Majīd",
+    meaning: "The Most Glorious",
+    category: "dhat",
+    root: "م-ج-د",
+    description: "Glory combining greatness with generosity — a majesty that is not remote or forbidding but radiates outward in gifts and honours upon creation.",
+  },
+  {
+    id: 51,
+    slug: "al-haqq",
+    arabic: "الْحَقّ",
+    transliteration: "Al-Haqq",
+    meaning: "The Truth",
+    category: "dhat",
+    root: "ح-ق-ق",
+    description: "The necessary, self-subsisting reality — the only being whose existence is not contingent; all else is real only insofar as it participates in His truth.",
+  },
+  {
+    id: 56,
+    slug: "al-hamid",
+    arabic: "الْحَمِيد",
+    transliteration: "Al-Hamīd",
+    meaning: "The All-Praiseworthy",
+    category: "dhat",
+    root: "ح-م-د",
+    description: "All praise in creation flows toward Him as to its natural source — praise that is not earned but belongs to the Essence as a necessary attribute.",
+  },
+  {
+    id: 63,
+    slug: "al-qayyum",
+    arabic: "الْقَيُّوم",
+    transliteration: "Al-Qayyūm",
+    meaning: "The Self-Subsisting",
+    category: "dhat",
+    root: "ق-و-م",
+    description: "He who subsists of Himself and on whom all else subsists — the Maturidi attribute of Qiyam bi-l-nafs (self-subsistence) embodied in this name.",
+  },
+  {
+    id: 64,
+    slug: "al-wajid",
+    arabic: "الْوَاجِد",
+    transliteration: "Al-Wājid",
+    meaning: "The Finder",
+    category: "dhat",
+    root: "و-ج-د",
+    description: "He who finds and perceives all things with no searching or seeking — a name of the Essence whose awareness precedes every created event.",
+  },
+  {
+    id: 65,
+    slug: "al-wahid",
+    arabic: "الْوَاحِد",
+    transliteration: "Al-Wāhid",
+    meaning: "The One",
+    category: "dhat",
+    root: "و-ح-د",
+    description: "Oneness in essence, attributes, and acts — the foundation of tawhid; no second being shares in or participates in this oneness.",
+  },
+  {
+    id: 66,
+    slug: "al-ahad",
+    arabic: "الْأَحَد",
+    transliteration: "Al-Ahad",
+    meaning: "The Unique",
+    category: "dhat",
+    root: "و-ح-د",
+    description: "Absolute singularity that admits no composition, division, or partner — the name that Sura Al-Ikhlas uses to negate every conceivable form of association.",
+  },
+  {
+    id: 67,
+    slug: "as-samad",
+    arabic: "الصَّمَد",
+    transliteration: "As-Samad",
+    meaning: "The Eternal",
+    category: "dhat",
+    root: "ص-م-د",
+    description: "The self-sufficient one to whom all creation turns in need — possessing no need of any kind while being the ultimate object of every creature's need.",
+  },
+  {
+    id: 72,
+    slug: "al-awwal",
+    arabic: "الْأَوَّل",
+    transliteration: "Al-Awwal",
+    meaning: "The First",
+    category: "dhat",
+    root: "أ-و-ل",
+    description: "He who precedes all things with no beginning — existence prior to time and before any created reality; the Qidam (eternity a parte ante) of the Essence.",
+  },
+  {
+    id: 73,
+    slug: "al-akhir",
+    arabic: "الْآخِر",
+    transliteration: "Al-Ākhir",
+    meaning: "The Last",
+    category: "dhat",
+    root: "أ-خ-ر",
+    description: "He who remains after all things — Baqa (eternal permanence); when all created things pass away, only the Divine Essence persists.",
+  },
+  {
+    id: 74,
+    slug: "az-zahir",
+    arabic: "الظَّاهِر",
+    transliteration: "Az-Zāhir",
+    meaning: "The Manifest",
+    category: "dhat",
+    root: "ظ-ه-ر",
+    description: "He who is manifest through His signs, effects, and clear proofs — more evident than any created thing, though not grasped by direct sensory perception.",
+  },
+  {
+    id: 75,
+    slug: "al-batin",
+    arabic: "الْبَاطِن",
+    transliteration: "Al-Bātin",
+    meaning: "The Hidden",
+    category: "dhat",
+    root: "ب-ط-ن",
+    description: "He who is hidden from direct comprehension — transcendence so complete that the Essence itself is beyond the reach of any created faculty.",
+  },
+  {
+    id: 83,
+    slug: "malik-ul-mulk",
+    arabic: "مَالِكُ الْمُلْك",
+    transliteration: "Mālik al-Mulk",
+    meaning: "Owner of All Sovereignty",
+    category: "dhat",
+    root: "م-ل-ك",
+    description: "Absolute ownership of every realm, domain, and kingdom — no sovereignty exists independently; all thrones are merely given in trust and retrieved at will.",
+  },
+  {
+    id: 84,
+    slug: "dhul-jalali-wal-ikram",
+    arabic: "ذُو الْجَلَالِ وَالْإِكْرَام",
+    transliteration: "Dhū al-Jalāl wa-l-Ikrām",
+    meaning: "Lord of Majesty and Bounty",
+    category: "dhat",
+    root: "ج-ل-ل",
+    description: "The pairing of transcendent awe-inspiring majesty (jalal) with active honouring generosity (ikram) — the complete divine character toward which all worship is directed.",
+  },
+  {
+    id: 87,
+    slug: "al-ghani",
+    arabic: "الْغَنِيّ",
+    transliteration: "Al-Ghanī",
+    meaning: "The Self-Sufficient",
+    category: "dhat",
+    root: "غ-ن-ي",
+    description: "Absolute richness and independence — needing nothing from creation while being the source from which all created needs are satisfied.",
+  },
+  {
+    id: 92,
+    slug: "an-nur",
+    arabic: "النُّور",
+    transliteration: "An-Nūr",
+    meaning: "The Light",
+    category: "dhat",
+    root: "ن-و-ر",
+    description: "The light of the heavens and the earth — not physical light but the ontological illumination by which anything can be known, seen, or guided.",
+  },
+  {
+    id: 95,
+    slug: "al-baqi",
+    arabic: "الْبَاقِي",
+    transliteration: "Al-Bāqī",
+    meaning: "The Everlasting",
+    category: "dhat",
+    root: "ب-ق-ي",
+    description: "Eternal permanence with no end — the Baqa attribute of the divine Essence; all created existence is borrowed and temporary while His is necessary and permanent.",
+  },
+  {
+    id: 99,
+    slug: "allah",
+    arabic: "اللَّه",
+    transliteration: "Allāh",
+    meaning: "The God",
+    category: "dhat",
+    root: "أ-ل-ه",
+    description: "The supreme proper name — Al-Ism al-A'zam (the Greatest Name) that encompasses all divine attributes; the name to which all other names are adjectives.",
+  },
+];
